Add putByte to Part and pre-allocate serialized form.

+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/BaseCommand.java
@@ -1461,7 +1461,7 @@ public abstract class BaseCommand implements Command {
       }
     }
 
-    return Operation.fromOrdinal((byte) operationPart.getInt());
+    return Operation.fromOrdinal(operationPart.getByte());
   }
 
 }
